Disable some tests that don't make sense for arm-nacl glibc
The version of the futex tests that tries to use libc internals
doesn't make sense with the new glibc, since it always relies on
the IRT futex calls.

The glibc_file64_test exists to test something that is not true
of the new glibc, and won't ever be.

The glibc_socket_wrappers exists to test the old x86 nacl-glibc's
ad hoc wrapping interfaces, which the new glibc will not support.

The glibc_static_test works fine with arm-nacl glibc just using
-static out of the box.

BUG= https://code.google.com/p/nativeclient/issues/detail?id=3068
TEST= trybots
R=phosek@chromium.org

Review URL: https://codereview.chromium.org/1115813003