Add --interp-prefix switch to elf_loader
In testing situations and in sel_ldr-based deployment situations,
the PT_INTERP name (e.g. "/lib/ld-nacl-arm.so.1") is not the file
name that should be opened directly.  It needs to be found under
the toolchain directory.  So give elf_loader a switch to prefix
the PT_INTERP file name with another directory name.

It turned out that the echo_interp test was not actually testing
the PT_INTERP handling at all.  The -dynamic-linker command-line
switch to the linker has no effect in what is actually a fully
static link, as is true of these test cases.  Instead, use the
alternate mechanism of the magic section name ".interp".

BUG= https://code.google.com/p/nativeclient/issues/detail?id=3068
TEST= trybots
R=dschuff@chromium.org, phosek@chromium.org

Review URL: https://codereview.chromium.org/1116243002