Do not serialize uselist order into pexes
Bitcode use lists cannot be parsed with lazy function materialization and
module splitting, which we use with debug pexes in the browser.
For now, disable writing them into the bitcode.

Also add a test that uses streaming and module splitting with nonstable bitcode.
In order to do that, change pnacl-translate to allow the command line flag to
override the default (which is no streaming in a few cases) and use streaming.
Add tests for the default behavior and the overriding behavior.

R=jvoung@chromium.org, mseaborn@chromium.org
BUG= https://code.google.com/p/nativeclient/issues/detail?id=4190

Review URL: https://codereview.chromium.org/1153283008