Making gsd uploads have a shorter overwrite window.
We've been seeing a recent uptick in colliding bucket writes.
This root cause is that google storage writes are not atomic.
In fact the -n (no clobber) option can be off by a wide margin
for large files that can take seconds to upload.
Instead of uploading directly, we will upload to a temporary name,
then copy into place and delete the temporary.
This will reduce, but not eliminate the possibility of collisions.

BUG=None
TEST=None
R=binji@chromium.org

Review URL: https://codereview.chromium.org/1278893002 .