Update chrome_rev
This updates chrome_rev and all the DEPS entries that should match it.

This includes the latest Chromium-built Clang, which implements new
warnings that bite in the PNaCl toolchain build (both in binutils and
in LLVM).  Disabling -Werror for those components' builds works around
that problem.

This also tweaks the toolchain_build rules for components built using
the Chromium-built Clang so that they consider the clang binaries as
inputs for memoization.  Without this, a DEPS change to bring in newer
clang binaries could be made without bots immediately attempting to
use the new clang in building the PNaCl toolchain components, and thus
keeping any problems hidden until some other change induced a cache miss.

BUG= none
TEST= trybots
R=bradnelson@chromium.org, dschuff@chromium.org

Review URL: https://codereview.chromium.org/1325603006 .