Add a test for x86-32 IRT stack realignment
This is a new test for an old bug.  It tries to ensure empirically that
the IRT code is pessimistically realigning the stack pointer to 16
bytes.  It does this by calling into the IRT with the stack
intentionally misaligned, and then checking that a callback from inside
that call reaches the callback function with a properly-aligned stack.
Later, this test will also run with an IRT built by GN and so verify
that the GN build of the IRT is using the right flags.

BUG= https://code.google.com/p/nativeclient/issues/detail?id=3935
TEST= trybots
R=dschuff@chromium.org

Review URL: https://codereview.chromium.org/1367083004 .