Update chrome_rev, clean up GN build
This updates chrome_rev and dependent entries in DEPS, including new
build/ stuff for GN.  With that, we can clean up the top-level GN build
script to match Chromium norms: ARM too is built with Clang; the NaCl
gcc-newlib toolchains are gone, nacl-clang-newlib is the new default.
Also clean up the GN build of nacl_helper_bootstrap, so it works with
Clang on ARM, and so it uses basic -O/-g flags appropriate for the
particular build type (roughly matching what Gyp does).

With all that, the GN build is now enabled on all Linux bots.

BUG= none
TEST= trybots
R=dpranke@chromium.org, bradnelson@chromium.org

Review URL: https://codereview.chromium.org/1361163003 .