GN: Fix invocation of win_as.py
The include directory needed is the top of the build, not the
native_client subdirectory.

The win_as.py script expects to find cl.exe in its PATH.
In the GYP build, this is accomplished by running the script
via gyp-win-tool action-wrapper, which is implicit in GYP.
In GN, there is no implicit wrapper, so use it explicitly.

BUG= 512869
R=bradnelson@chromium.org, dpranke@chromium.org, scottmg@chromium.org

Review URL: https://codereview.chromium.org/1414453014 .