Modify spec2k run script so it doesn't blow away valuable Subzero files.
The prepare_input script for 176.gcc removes all .s files before running the benchmark.  This is inconvenient for Subzero development and testing because Subzero's .s output for the gcc benchmark gets deleted as well.

To help with this, we can specify a blacklist pattern of files that should't be deleted, even if they are listed for removal in the config or on the command line.

BUG= none
R=kschimpf@google.com

Review URL: https://codereview.chromium.org/1494473007 .