Make FYI bot test script exit with nonzero status if any tests fail.
The PNaCl FYI bots are getting errors running torture and LLVM testsuite
tests but the bots are showing as green. I suspect that previously the
build waterfall would show the bot as red if any step was red (i.e. the
script emitted @@@STEP_FAILURE@@@ but at some point it changed to be
based on the exit status of the script.

BUG=green bots
R=mseaborn@chromium.org, stichnot@chromium.org

Review URL: https://codereview.chromium.org/1700503002 .