Scons cleanup: Stop DualLibrary() from compiling source files twice on Linux
DualLibrary() is left over from when we built NPAPI/PPAPI plugins and
wanted to compile some libraries both with and without -fPIC.  We
don't need this behaviour any more -- the "*_shared.a" library
variants are not used.

This should reduce the amount of work the bots do.

BUG=none
TEST=trybots

Review URL: https://codereview.chromium.org/1737153002