Remove GYP build from bots and 'gclient runhooks'
This starts the process of eliminating the GYP build of standalone NaCl.
This doesn't touch any .gyp or .gypi files, most of which are still used
in the Chromium build (which still needs to support GYP for a little
while longer).  It removes the GYP-related steps from the bot runs, and
removes the machinery for running GYP manually or via gclient.

BUG= none
TEST= trybots
R=dschuff@chromium.org

Review URL: https://codereview.chromium.org/1924743004 .