Remove host libcxx from toolchain build
This dates back to when LLVM required a newer C++ stdlib than was
available on our bots; we used a custom-built libcxx instead of
libstdc++.  Now libc++ is the default on mac builds and the libstdc++
in our sysroot is good enough to build LLVM, this workaround is no
longer needed.

Aside from removing the host libcxx headers and libs, it also works
around a bug in le32-nacl-clang's header include path handling that
caused it to use the host copy of the libcxx headers (it just makes
an extra copy of the target's libcxx headers, which are the same).

Bug: https://bugs.chromium.org/p/nativeclient/issues/detail?id=4421
Change-Id: Idb4edf2949d2c8691833fa8389449cf9b275aea4
Reviewed-on: https://chromium-review.googlesource.com/c/native_client/src/native_client/+/2093138
Commit-Queue: Derek Schuff <dschuff@chromium.org>
Reviewed-by: Fabian Sommer <fabiansommer@chromium.org>