Use glibc compiler instead of gcc newlib toolchain to build libgcc_eh_x86_64
We use libgcc_eh for unwinding support in the clang-based compilers, and
gcc is required to compile it. Currently we are using gcc from the old
newlib toolchain, but we do not have a mac x86-64 package for it, and
don't want to build one. We do have a mac x86-64 package for the glibc
toolchain though, so this CL switches to that for building libgcc_eh.
Since the compiler binaries are the same and libgcc_eh doesn't use
anything from libc, the output is identical.

Change-Id: I1d27ebf1c6d43f8cef0a0de05587527b21065f47
Reviewed-on: https://chromium-review.googlesource.com/c/native_client/src/native_client/+/2857368
Reviewed-by: Fabian Sommer <fabiansommer@chromium.org>