Adapt debugger_test.c for saigo on ARM
Change the inline assembly that is part of the debugger test. On saigo,
the branch instruction will get autosandboxed, so the inline assembly
itself does not need to apply any sandboxing to it.

If the inline assembly does apply needless sandboxing, it confuses the
test expectation which expects a bic + branch in this place, but finds
two bics + a branch instead.

Bug: 1106311
Change-Id: I6cfd5c9ed1105855ce20d6e0745a6046c3306ce8
Reviewed-on: https://chromium-review.googlesource.com/c/native_client/src/native_client/+/3097726
Reviewed-by: Derek Schuff <dschuff@chromium.org>
Commit-Queue: Fabian Sommer <fabiansommer@chromium.org>