[Py3] Validator trie comparisons: Fix bug on Py2 and incompatibility on Py3
On Python 2, "!=" does not automatically forward to the __eq__ method.
That causes a minor bug where DiffTries() can report differences
between tries when there aren't any, when they contain AcceptInfos
with different identities but the same contents.  Fix this by adding
an __ne__ implementation and add a test.  I suspect this bug didn't
manifest itself in practice because tries would be interned via the
NodeCache mechanism and so AcceptInfos would be de-duplicated.

On Python 3, "!=" does automatically forward to the __eq__ method,
causing a different problem which is caught by the existing
testTrieDiff test.  The problem arises with the comparison
"node1.accept_info != node2.accept_info" in DiffTries() when
node1.accept_info is not None and node2.accept_info is None.  Fix by
adding an "is None" check.  Note that if node1.accept_info is None
there is no problem, because the comparison won't be forwarded to
AcceptInfo's __eq__ method.

Bug: 1205898
Change-Id: I238195270f98502ab9a76754a684ae542c56873a
Reviewed-on: https://chromium-review.googlesource.com/c/native_client/src/native_client/+/3206740
Reviewed-by: Fabian Sommer <fabiansommer@chromium.org>
Commit-Queue: Mark Seaborn <mseaborn@chromium.org>