Remove unused #include of nacl_log.h in nacl_error_code.c
This was necessary for a proof-of-concept of building ARM32 NaCl in
ARM64 Chromium, because nacl_log.h #includes portability.h, which
complains about unrecognised architectures such as ARM64.

This might turn out not to be necessary or sufficient overall, but
removing unused #includes is generally good to do anyway.

Bug: chromium:1299021
Change-Id: Ib4123169c5d08671cae0c64cc6281bd864d8c8c8
Reviewed-on: https://chromium-review.googlesource.com/c/native_client/src/native_client/+/3473401
Reviewed-by: Derek Schuff <dschuff@chromium.org>
Commit-Queue: Mark Seaborn <mseaborn@chromium.org>