Roll build deps
This rolls in a change required for our new Mac12 bots from build/

Rename a clang-format file that got renamed for Mac
Changes to src/trusted to fix Wdeclaration-after-statement warnings on
Mac.
There is a new sysroot with a new name (sid -> bullseye).
Since this name is baked into the arm trusted toolchain, roll a new
toolchain with that change.

Bug: 1284673
Change-Id: I26424af3dee95e372b6df26c9136920515bca545
Reviewed-on: https://chromium-review.googlesource.com/c/native_client/src/native_client/+/3755486
Commit-Queue: Fabian Sommer <fabiansommer@chromium.org>
Reviewed-by: Mark Seaborn <mseaborn@chromium.org>