Check in and use old Googletest
Check in a copy of the 2013-vintage Googletest which is specified in the
DEPS file. Trying to use an external one would not work because an old
version that supports old C++ standards is needed. Hook it up to the
build system so that most tests work now.

Imported from gtest revision 2d3543f81d6d4583332f8b60768ade18e0f96220:
- LICENSE
- src/
- *.h files within include/