Fix REGS_SAVER_FUNC test util on ARM
REGS_SAVER_FUNC is meant to work even if the stack pointer is
misaligned, but the multi-register PUSH instructions crash if the stack
is misaligned, as per the ARM documentation. I guess the old(?) bundled
QEMU used by upstream doesn't implement this faithfully. Fix it by using
a series of single-register pushes instead.

Fixes the run_register_set_test test target on ARM.